% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/becker09.r
\docType{data}
\name{becker09}
\alias{becker09}
\title{Becker 2009 data}
\format{A dataframe with 10 rows and 9 columns:
\describe{
   \item{ID}{Study ID}
   \item{N}{Study sample size}
   \item{Team}{...}
   \item{Cognitive_Performance}{Correlation between Cognitive and Performance}
   \item{Somatic_Performance}{Correlation between Somatic and Performance}
   \item{Selfconfidence_Performance}{Correlation between Self Confidence and
   Performance}
   \item{Somatic_Cognitive}{Correlation between Somatic and Cognitive}
   \item{Selfconfidence_Cognitive}{Correlation between Self Confidence and
   Cognitive}
   \item{Selfconfidence_Somatic}{Correlation between Self Confidence and
   Somatic}
   }}
\usage{
becker09
}
\description{
This data contains data originally collected by Becker in
 2009 on ...
}
\examples{
  becker09
}
\keyword{datasets}
