% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.r
\name{c_mat_ft}
\alias{c_mat_ft}
\title{c mat ft}
\usage{
c_mat_ft(model_input, R)
}
\arguments{
\item{model_input}{Model input as a character string. Multiple models need
to be on their own line. Model syntax uses lavann like syntax, see details
for more details about this syntax.}

\item{R}{A correlation matrix, most likely this will be the average
correlation matrix outputted from the metafor package.}
}
\value{
A list.
}
\description{
c mat ft
}
\examples{
Br <-  matrix(c(1.00000000, -0.09773331, -0.1755029,  0.3186775,
-0.09773331,  1.00000000,  0.5271873, -0.4175596,
-0.17550292,  0.5271872,  1.0000000, -0.4006848,
0.31867753, -0.41755963, -0.4006848,  1.0000000),
nrow = 4, byrow = TRUE)

colnames(Br) <- c("Performance",  "Self_confidence",  "Cognitive", "Somatic" )

rownames(Br) <- colnames(Br)

## Proposed path model
model <- "## Regression paths
Performance ~  Self_confidence  + Cognitive  + Somatic
Self_confidence ~ Cognitive + Somatic "

c_mat_ft(model, Br)
}
