% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_meta.r
\name{corr_meta}
\alias{corr_meta}
\title{Master function}
\usage{
corr_meta(data, n, type = c("average", "weighted", "simple"),
  missing = "remove", variable_names = NULL, ID = NULL, effect_size,
  var_cor, weights = NULL, moderators = NULL, random_params = NULL,
  structure = "CS", test = "t", intercept = FALSE,
  estimation_method = "REML", model = NULL, num_obs = NULL, ...)
}
\arguments{
\item{data}{Raw data input}

\item{n}{Sample size}

\item{type}{type of adjustment (olkin siotani)}

\item{missing}{What to do with missing data}

\item{variable_names}{An optional character vector specifying variable names
of original data. This is mostly used to pass to \code{\link{df_to_corr}}.}

\item{ID}{A variable name, as a character string, to use as names for the list elements.}

\item{model}{This is model syntax specified in the format by lavaan. See
\code{\link{sem}} for more details about model syntax.}

\item{num_obs}{Number of observations}

\item{...}{Additional parameters passed to metafor. See \code{\link{rma.mv}}
for more details.}
}
\description{
This function does all of the steps in a single unified function.
}
