% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_model.r
\name{extract_model}
\alias{extract_model}
\title{Extract relavent model values from metafor}
\usage{
extract_model(model)
}
\arguments{
\item{model}{A metafor model object, likely coming from
\code{\link{fit_model}}.}
}
\description{
Extract relavent model values from metafor
}
