% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.r
\name{find_reg_coef}
\alias{find_reg_coef}
\title{Find regression coefficients}
\usage{
find_reg_coef(model_input, R, ...)
}
\arguments{
\item{model_input}{Model input as a character string. Multiple models need
to be on their own line. Model syntax uses lavann like syntax, see details
for more details about this syntax.}

\item{R}{A correlation matrix, most likely this will be the average
correlation matrix outputted from the metafor package.}

\item{...}{Currently not used}
}
\value{
A vector of regression coefficient estimates.
}
\description{
Find regression coefficients
}
