% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.r
\name{fit_model}
\alias{fit_model}
\title{Fit meta analysis models with metafor}
\usage{
fit_model(data, effect_size, var_cor, weights = NULL,
  moderators = NULL, random_params = NULL, structure = "UN",
  test = "t", intercept = FALSE, estimation_method = "REML", ...)
}
\arguments{
\item{data}{Raw data input}

\item{effect_size}{A vector of observed effect sizes to be modeled with
metafor. See \code{\link{rma.mv}} for more details.}

\item{var_cor}{Variance covariance matrix of sampling errors or sampling weights
passed on to metafor. See \code{\link{rma.mv}} for more details.}

\item{weights}{Optional weights matrix passed to metafor. Default is NULL.
See \code{\link{rma.mv}} for more details.}

\item{moderators}{Moderators (covariates, predictors) to be included in the
analysis with metafor. Default is NULL meaning no predictors.
See \code{\link{rma.mv}} for more details.}

\item{random_params}{Specification of the random effect structure. This
argument takes the form a single sided formula. Default is NULL which
means a fixed effects model will be fitted with metafor.
See \code{\link{rma.mv}} for more details.}

\item{structure}{The variance structure passed to metafor. Default is "CS"
which represents compound symmetry. See \code{\link{rma.mv}} for more
details.}

\item{test}{What type of test statistic should be used.
Default is a t-statistic. See \code{\link{rma.mv}} for more details.}

\item{intercept}{Should an intercept be included in the metafor model.
See \code{\link{rma.mv}} for more details.}

\item{estimation_method}{Estimation method to pass on to metafor. Default is
"REML". See \code{\link{rma.mv}} for more details.}

\item{...}{Additional parameters passed to metafor. See \code{\link{rma.mv}}
for more details.}
}
\description{
This function is a wrapper that calls metafor to fit models.
}
