% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jacobian.r
\name{jacobian_pcor}
\alias{jacobian_pcor}
\title{Variance-covariance Matrix of Partials}
\usage{
jacobian_pcor(R)
}
\arguments{
\item{R}{Matrix of estimates}
}
\description{
Variance-covariance Matrix of Partials
}
