% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_correlations.r
\docType{data}
\name{math_correlations}
\alias{math_correlations}
\title{Math Correlations Data}
\format{A nested list with correlations and sample size:
\describe{
   \item{math_correlations}{A list of math correlations}
   \item{sample_size}{Study sample size}
   }}
\usage{
math_correlations
}
\description{
This data contains correlations on math ...
}
\examples{
  math_correlations
}
\keyword{datasets}
