% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.r
\name{model_fit}
\alias{model_fit}
\title{Model fitting function}
\usage{
model_fit(model_result, R, method_mat = "lavaan", method_null = "sem",
  N)
}
\arguments{
\item{R}{Correlation matrix}

\item{N}{}
}
\description{
Model fitting function
}
