% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OS.r
\name{olkin_siotani}
\alias{olkin_siotani}
\title{Olkin & Siotani 1976 variance-covariance matrix}
\usage{
olkin_siotani(data, n, type = c("average", "weighted", "simple"))
}
\arguments{
\item{data}{A correlation matrix or a list of correlation matrices.}

\item{n}{Sample size}

\item{type}{The type of Olkin and Siotani correction to make.}
}
\description{
Description here
}
