% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.r
\name{prep_data}
\alias{prep_data}
\title{Prep Data for Metafor}
\usage{
prep_data(data, n, type = c("average", "weighted", "simple"),
  missing = "remove", variable_names = NULL)
}
\arguments{
\item{n}{Sample size}

\item{variable_names}{An optional character vector specifying variable names
of original data. This is mostly used to pass to \code{\link{df_to_corr}}.}
}
\description{
Prep Data for Metafor
}
