% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_meta.r
\name{corr_meta}
\alias{corr_meta}
\title{Master function}
\usage{
corr_meta(
  data,
  n,
  type = c("average", "weighted", "simple"),
  missing = "remove",
  variable_names = NULL,
  ID = NULL,
  effect_size,
  var_cor,
  weights = NULL,
  moderators = NULL,
  random_params = NULL,
  structure = "CS",
  test = "t",
  intercept = FALSE,
  estimation_method = "REML",
  model = NULL,
  num_obs = NULL,
  ...
)
}
\arguments{
\item{data}{Raw data input}

\item{n}{Sample size}

\item{type}{type of adjustment (olkin siotani)}

\item{missing}{What to do with missing data}

\item{variable_names}{An optional character vector specifying variable names
 of original data. This is mostly used to pass to \code{\link{df_to_corr}}.
 A character vector of variable names representing the columns
to convert to a pairwise correlation matrix. Variable names take the following format:
"columnname_rowname" where the name prior to the underscore is the column name and the
name after the underscore is the row name.}

\item{ID}{A variable name, as a character string, to use as names for the list elements.}

\item{effect_size}{A vector of observed effect sizes to be modeled with
metafor. See \code{\link{rma.mv}} for more details.}

\item{var_cor}{Variance covariance matrix of sampling errors or sampling weights
passed on to metafor. See \code{\link{rma.mv}} for more details.}

\item{weights}{Optional weights matrix passed to metafor. Default is NULL.
See \code{\link{rma.mv}} for more details.}

\item{moderators}{Moderators (covariates, predictors) to be included in the
analysis with metafor. Default is NULL meaning no predictors.
See \code{\link{rma.mv}} for more details.}

\item{random_params}{Specification of the random effect structure. This
argument takes the form a single sided formula. Default is NULL which
means a fixed effects model will be fitted with metafor.
See \code{\link{rma.mv}} for more details.}

\item{structure}{The variance structure passed to metafor. Default is "CS"
which represents compound symmetry. See \code{\link{rma.mv}} for more
details.}

\item{test}{What type of test statistic should be used.
Default is a t-statistic. See \code{\link{rma.mv}} for more details.}

\item{intercept}{Should an intercept be included in the metafor model.
See \code{\link{rma.mv}} for more details.}

\item{estimation_method}{Estimation method to pass on to metafor. Default is
"REML". See \code{\link{rma.mv}} for more details.}

\item{model}{This is model syntax specified in the format by lavaan.}

\item{num_obs}{Number of observations}

\item{...}{Additional parameters passed to metafor. See \code{\link{rma.mv}}
for more details.}
}
\description{
This function does all of the steps in a single unified function.
}
\seealso{
See \url{https://lavaan.ugent.be/} for more information on lavaan
syntax.
}
