% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.r
\name{fit_model}
\alias{fit_model}
\title{Fit meta analysis models with metafor}
\usage{
fit_model(data, effect_size, var_cor, weights = NULL, moderators = NULL,
  random_params = NULL, structure = "CS", test = "t", intercept = FALSE,
  estimation_method = "REML", ...)
}
\arguments{
\item{...}{Additional parameters passed to metafor. See \code{\link{rma.mv}}
for more details.}
}
\description{
This function is a wrapper that calls metafor to fit models.
}
