% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimization.r
\name{model_fit}
\alias{model_fit}
\title{Model fitting function}
\usage{
model_fit(model_input, R, method_mat = "loehlin", method_null = "sem",
  N)
}
\arguments{
\item{model_input}{Model input as a character string. Multiple models need
to be on their own line. Model syntax uses lavann like syntax, see details
for more details about this syntax.}

\item{R}{A correlation matrix, most likely this will be the average
correlation matrix outputted from the metafor package.}

\item{method_mat}{Method of estimation, can either be "loehlin" or "lavaan".
Default is "loehlin"}

\item{method_null}{Unsure}

\item{N}{Sample size}
}
\value{
A list of fit indices.
}
\description{
Model fitting function
}
\examples{
Br <-  matrix(c(1.00000000, -0.09773331, -0.1755029,  0.3186775,
-0.09773331,  1.00000000,  0.5271873, -0.4175596,
-0.17550292,  0.5271872,  1.0000000, -0.4006848,
0.31867753, -0.41755963, -0.4006848,  1.0000000),
nrow = 4, byrow = TRUE)

colnames(Br) <- c("Performance",  "Self_confidence",  "Cognitive", "Somatic" )

rownames(Br) <- colnames(Br)

## Proposed path model
model <- "## Regression paths
Performance ~  Self_confidence  + Cognitive  + Somatic
Self_confidence ~ Cognitive + Somatic "

N <- 573
model_fit(model_input = model, R = Br, method_mat  = "lavaan",
         method_null = "sem", N)
model_fit(model_input = model, R = Br, method_mat  = "loehlin",
         method_null = "sem", N )
}
