% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.r
\name{path_model}
\alias{path_model}
\title{Path Model Function}
\usage{
path_model(data, model, num_obs, adjust_se = TRUE, ...)
}
\arguments{
\item{data}{A list that contains the correlation matrix for model fitting
and the variance matrix. This would most likely come from the
\code{\link{extract_model}} function.}

\item{model}{This is model syntax specified in the format by lavaan. See
\code{\link{sem}} for more details about model syntax.}

\item{num_obs}{Number of observations}

\item{adjust_se}{Adjust the standard errors to reflect the ...}

\item{...}{Additional options to pass to \code{\link{sem}}.}
}
\description{
This function fits the path model and returns adjusted standard errors.
}
